fix bug observed when broker cluster has no topics
Don't treat it as an error condition if the broker cluster has no
topics, since this is the initial state of a newly provisioned
cluster.  Treating this condition as an error was preventing the
router thread from responding to requests to automatically create a
topic.

Additionally, do some minor cleanup and fixing of the metadata
handling code.