add gzip compression support, and almost finished lz4 support
To finish LZ4 support, dory's wire protocol implementation needs to be
updated to support asking the brokers what version of Kafka they are
running.  LZ4 support will be enabled only for broker versions >=
0.10.0.0.  Enabling LZ4 for earlier versions would require a messy
workaround for a bug in Kafka.  See
https://cwiki.apache.org/confluence/display/KAFKA/KIP-57+-+Interoperable+LZ4+Framing
for details.

The LZ4 implementation provided here has been tested, and verified to
work with Kafka version 0.10.0.1.  To enable it, look for the following
line of code in src/dory/dory.cc:

              bool enable_lz4 = false;

and set the value to true.  Just be sure to use it only with Kafka
versions >= 0.10.0.0.  Bad things will happen if used with earlier
versions.