fix bug in TEventSemaphore::Reset(): let DTOR close new_fd
This prevents the destructor for new_fd from trying to close an
already closed file descriptor, by avoiding an unnecessary manual
close() outside the destructor.  The bug may never cause problems
in practice, but it's possible that the file descriptor has already
been reallocated by a different thread by the time the new_fd
destructor closes it a second time, potentially causing the newly
allocated file descriptor to be closed by mistake.  This would
likely be a rare but hard to track down problem.